+++ package.json
@@ -9,6 +9,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-ga": "^2.7.0",
+    "react-query": "^3.34.16",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.1",
     "sass": "^1.49.7"
