+++ src/App.js
@@ -1,21 +1,23 @@
-import React, { useRef, useEffect } from 'react';
-import { useLocation, Switch } from 'react-router-dom';
-import AppRoute from './utils/AppRoute';
-import ScrollReveal from './utils/ScrollReveal';
+import React, { useRef, useEffect } from "react";
+import { useLocation, Switch } from "react-router-dom";
+import AppRoute from "./utils/AppRoute";
+import ScrollReveal from "./utils/ScrollReveal";
+import { QueryClient, QueryClientProvider } from "react-query";
 
 // Layouts
-import LayoutDefault from './layouts/LayoutDefault';
+import LayoutDefault from "./layouts/LayoutDefault";
 
-// Views 
-import Home from './views/Home';
+// Views
+import Home from "./views/Home";
+
+const queryClient = new QueryClient();
 
 const App = () => {
-
   const childRef = useRef();
   let location = useLocation();
 
   useEffect(() => {
-    document.body.classList.add('is-loaded')
+    document.body.classList.add("is-loaded");
     childRef.current.init();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [location]);
@@ -25,10 +27,13 @@ const App = () => {
       ref={childRef}
       children={() => (
         <Switch>
-          <AppRoute exact path="/" component={Home} layout={LayoutDefault} />
+          <QueryClientProvider client={queryClient}>
+            <AppRoute exact path="/" component={Home} layout={LayoutDefault} />
+          </QueryClientProvider>
         </Switch>
-      )} />
+      )}
+    />
   );
-}
+};
 
 export default App;
