+++ src/components/sections/GenericSection.js
@@ -1,18 +1,19 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import SectionHeader from './partials/SectionHeader';
-import { SectionProps } from '../../utils/SectionProps';
+import React from "react";
+import PropTypes from "prop-types";
+import classNames from "classnames";
+import SectionHeader from "./partials/SectionHeader";
+import { SectionProps } from "../../utils/SectionProps";
+import Achievements from "../../Achievements";
 
 const propTypes = {
   children: PropTypes.node,
-  ...SectionProps.types
-}
+  ...SectionProps.types,
+};
 
 const defaultProps = {
   children: null,
-  ...SectionProps.defaults
-}
+  ...SectionProps.defaults,
+};
 
 const GenericSection = ({
   className,
@@ -25,40 +26,37 @@ const GenericSection = ({
   invertColor,
   ...props
 }) => {
-
   const outerClasses = classNames(
-    'section',
-    topOuterDivider && 'has-top-divider',
-    bottomOuterDivider && 'has-bottom-divider',
-    hasBgColor && 'has-bg-color',
-    invertColor && 'invert-color',
+    "section",
+    topOuterDivider && "has-top-divider",
+    bottomOuterDivider && "has-bottom-divider",
+    hasBgColor && "has-bg-color",
+    invertColor && "invert-color",
     className
   );
 
   const innerClasses = classNames(
-    'section-inner pt-0',
-    topDivider && 'has-top-divider',
-    bottomDivider && 'has-bottom-divider'
+    "section-inner pt-0",
+    topDivider && "has-top-divider",
+    bottomDivider && "has-bottom-divider"
   );
 
   const sectionHeader = {
-    title: 'Game Title',
-    paragraph: 'Achievements for <Game Title>'
+    title: "Game Title",
+    paragraph: "Achievements for <Game Title>",
   };
 
   return (
-    <section
-      {...props}
-      className={outerClasses}
-    >
+    <section {...props} className={outerClasses}>
       <div className="container">
         <div className={innerClasses}>
           <SectionHeader data={sectionHeader} className="center-content" />
+          <Achievements></Achievements>
         </div>
       </div>
     </section>
   );
-}
+};
 
 GenericSection.propTypes = propTypes;
 GenericSection.defaultProps = defaultProps;
